refactor: Create dedicated package for slice utilities
