+++ pkg/cslices/subslice.go
@@ -1,17 +1,6 @@
-package utils
+package cslices
 
 import "slices"
-
-// Remove an element from an integer slice by index.
-//
-// If the index is out of bounds, the original slice is returned.
-func RemoveInt(slice []int, index int) []int {
-	if index < 0 || index >= len(slice) {
-		return slice
-	}
-	result := append([]int{}, slice...) // Copy the slice to avoid modifying the original.
-	return append(result[:index], result[index+1:]...)
-}
 
 // Get the indexes of the first of a and its' closest occurance of b within
 // slice s.
